moving databse.js to utils folder
