+++ src/routes.js
@@ -1,6 +1,6 @@
 import { randomUUID } from 'node:crypto'
 import { buildRoutePath } from "./utils/build-route-path.js"
-import { Database } from './middlewares/database.js'
+import { Database } from './utils/database.js'
 
 const database = new Database()
